Reduce net consumption (#35)
* main: increase update timeout loop

In order to reduce the birdge-marker net consumtion, This commit reduces the number
of API calls by increasing the default poll-intervall.

Signed-off-by: alonsadan <asadan@redhat.com>

* reported resources: use cache

In order to reduce the bridge-marker net consumption, this commit reduces the
number of API calls by adding a cache to keep track of the bridges already been
reported, and update the node object only every 10 minutes.

Signed-off-by: alonsadan <asadan@redhat.com>

* main: add jitter to update & reconcile

To avoid communication peeks, this commit adds a jitter to randomize
the duration it takes to update and retrieve info from the node.

Signed-off-by: alonsadan <asadan@redhat.com>