marker: take cache pointer for Update
The cache should be passed as pointer and not by value, or otherwise
the updates on the cache done in Update() won't be actually reflected on the
"real" cache.

This commit updates the main loop and the marker to use the cache as a pointer.
Also it changes the the cache var name, to be different than the type name,
because this could be confusing.

Signed-off-by: alonsadan <asadan@redhat.com>