Pass cache as pointer (#38)
* marker_test: increase timeout

commit 743b3f9f97c9ebca01c0ec9f39025f47ab583d93
increased the update interval of the bridge marker from 10 to 60 seconds,
and the tests should have been updated accordingly. This commit updates the
tests timeout from 20 to 180 in accordance with the update interval increasing.

Signed-off-by: alonsadan <asadan@redhat.com>

* marker: take cache pointer for Update

The cache should be passed as pointer and not by value, or otherwise
the updates on the cache done in Update() won't be actually reflected on the
"real" cache.

This commit updates the main loop and the marker to use the cache as a pointer.
Also it changes the the cache var name, to be different than the type name,
because this could be confusing.

Signed-off-by: alonsadan <asadan@redhat.com>

* cluster-sync: pull image from registry

When making changes locally, and running `make cluster-sync` the changes did
not take place because the image for the marker pod was pulled from quay.
This commit makes it take it from the local registry.

Signed-off-by: alonsadan <asadan@redhat.com>