tests, marker-test: increase timeout
Recently the update interval of the node  was increased to 1 minute, and also a jitter
with 1.2 factor was added. This means that if for some reason one update was missed,
in the worst-case scenario there could be almost 2 minutes and a half difference between updates.

This could  makes the 120 interval not enough for the a pod to turn from Pending to Running
based on the label added to the node.

Therefor this commit increases the timeout for 240 seconds which should be enough for
the label to appear on the node and for the pod to become ready.

Signed-off-by: alonsadan <asadan@redhat.com>